pyproject.toml
[project]
name = "bsky-jetstream"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "James Thigpen", email = "james@yesawoo.com" }]
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "dnspython>=2.7.0",
    "httpx>=0.28.1",
    "httpx-ws>=0.7.2",
    "pytest>=8.4.1",
    "zstandard>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jetstream"]   

[project.scripts]
jetstream = "jetstream.jetstream:jetstream"
